Refatorando e organizando codigo com classes
