+++ semana19/labBook/src/routes/userRoute.ts
@@ -1,10 +1,10 @@
 import express from "express";
-import signUp from '../controller/signUp'
-import signIn from '../controller/signIn'
+import _UserController from "../controller/userController";
 
 const userRoute = express.Router();
+const UserController = new _UserController();
 
-userRoute.post("/signup", signUp)
-userRoute.post("/signin", signIn)
+userRoute.post("/signup", UserController.signUp);
+userRoute.post("/signin", UserController.signIn);
 
 export default userRoute;
